Refactoring Icicle Graph into smaller components
