+++ ui/packages/shared/profile/src/index.tsx
@@ -18,7 +18,7 @@ import ProfileTypeSelector from './ProfileTypeSelector';
 import type {FlamegraphData, TopTableData} from './ProfileView';
 import {QueryServiceClient} from '@parca/client';
 
-export * from './IcicleGraph';
+export * from './ProfileIcicleGraph/IcicleGraph';
 export * from './ProfileIcicleGraph';
 export * from './ProfileSource';
 export * from './ProfileView';
